+++ packages/module-ag-grid/src/module.ts
@@ -1,6 +1,11 @@
 import { IAgGridConfigurator, AgGridConfigurator } from './configurator';
 import { IAgGridProvider, AgGridProvider } from './provider';
-import type { Module, ModulesInstanceType } from '@equinor/fusion-framework-module';
+import type {
+    IModuleConfigurator,
+    IModulesConfigurator,
+    Module,
+    ModulesInstanceType,
+} from '@equinor/fusion-framework-module';
 
 export type AgGridModule = Module<'agGrid', IAgGridProvider, IAgGridConfigurator>;
 
@@ -14,6 +19,19 @@ export const module: AgGridModule = {
         return config;
     },
     initialize: ({ config }): IAgGridProvider => new AgGridProvider(config),
+};
+
+export const configureAgGrid = (args: {
+    licenseKey: string;
+}): IModuleConfigurator<AgGridModule> => ({
+    module,
+    configure: (config) => {
+        config.licenseKey = args.licenseKey;
+    },
+});
+
+export const enableAgGrid = (config: IModulesConfigurator): void => {
+    config.addConfig({ module });
 };
 
 export default module;