+++ packages/app/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@1.0.1...@equinor/fusion-framework-app@1.0.2) (2022-09-13)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## [1.0.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@1.0.1-next.1...@equinor/fusion-framework-app@1.0.1) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## [1.0.1-next.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@1.0.1-next.0...@equinor/fusion-framework-app@1.0.1-next.1) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## [1.0.1-next.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@1.0.0...@equinor/fusion-framework-app@1.0.1-next.0) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## 1.0.0 (2022-09-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* **app:** config is now object
+
+### Features
+
+* **app:** update init ([528d77f](https://github.com/equinor/fusion-framework/commit/528d77f30430242d05bf9a7fbed3ea68171df39d))
+
+
+### Bug Fixes
+
+* **app:** expect fusion modules ([301fcab](https://github.com/equinor/fusion-framework/commit/301fcab23e857bf87440b4212513a7eea2641aea))
+
+
+
+## [1.0.0-alpha.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-app@0.2.3...@equinor/fusion-framework-app@1.0.0-alpha.0) (2022-09-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* **app:** config is now object
+
+### Features
+
+* **app:** update init ([528d77f](https://github.com/equinor/fusion-framework/commit/528d77f30430242d05bf9a7fbed3ea68171df39d))
+
+
+### Bug Fixes
+
+* **app:** expect fusion modules ([301fcab](https://github.com/equinor/fusion-framework/commit/301fcab23e857bf87440b4212513a7eea2641aea))
+
+
+
+## 0.2.3 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
+## 0.2.2 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-app
+
+
+
+
+
 ## 0.2.1 (2022-09-01)
 
 **Note:** Version bump only for package @equinor/fusion-framework-app