+++ packages/react-module-event/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@1.0.1...@equinor/fusion-framework-react-module-event@1.0.2) (2022-09-13)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [1.0.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@1.0.1-next.1...@equinor/fusion-framework-react-module-event@1.0.1) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [1.0.1-next.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@1.0.1-next.0...@equinor/fusion-framework-react-module-event@1.0.1-next.1) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [1.0.1-next.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@1.0.0...@equinor/fusion-framework-react-module-event@1.0.1-next.0) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## 1.0.0 (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [1.0.0-alpha.0](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@0.2.3...@equinor/fusion-framework-react-module-event@1.0.0-alpha.0) (2022-09-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [0.2.3](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@0.2.2...@equinor/fusion-framework-react-module-event@0.2.3) (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## [0.2.2](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-module-event@0.2.1...@equinor/fusion-framework-react-module-event@0.2.2) (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
+## 0.2.1 (2022-09-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-module-event
+
+
+
+
+
 # 0.2.0 (2022-08-23)